Get BPM from a possible VST host
This change requires also moving setDelayBufferParams to a separate thread.
This is done because playhead object used for getting the BPM can only be
obtained in the processBlock-function, and to keep the signal processing
quick the BPM parameter needs to be set in a separate thread. There are
added WaitableEvent to keep things in sync, and most likely fix earlier race
condition as well